Create TeamMeetingLogs
he minutes of the meeting with your TA every week will be recorded in a file here. You will be updating the same file every week with the latest updates. Points to include in the minutes would be:
Decisions made
Alternative actions/options discussed
Follow-up items, including agreed-on roles and responsibilities